Kanguva (𝟸𝟶𝟸𝟺) 𝙵𝚞𝚕𝚕 𝙼𝚘𝚟𝚒𝚎 𝙳𝚘𝚠𝚗𝚕𝚘𝚊𝚍 𝚃𝚊𝚖𝚒𝚕 + H𝚒𝚗𝚍𝚒
"Kanguva" is a highly anticipated Indian historical action-adventure film directed by Siva and starring Suriya in the lead role. Set to release in 2024, the film has generated a lot of buzz due to its unique plot, which is rumored to blend elements of mythology and historical drama with high-octane action sequences.